Fix annotations on False Positive/False Negative Performance Plot (#284)
The individual model view has a chart showing false positive and false negative rates (defined in `PerformancePlot.tsx`). The axis of this plot was in percent, but the annotations on the bars themselves were decimal. This problem was due to the incorrect key to `PerformanceMetrics.ts` being passed into `FormatMetrics.ts` by `PerformancePlot.tsx` - in the `PerformanceMetric` map, the names `fallout_rate` and `miss_rate` are used. This fixes that mismatch.

Had to force in because for some reason the successful builds were not registered.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>