sample notebook changes and readme (#282)
* sample notebook changes and readme first draft

* addressing PR comments and adding EA screenshots

* bump requirement version of rai_core_flask in raiwidgets (#283)

Signed-off-by: Roman Lutz <rolutz@microsoft.com>

* Fix annotations on False Positive/False Negative Performance Plot (#284)

The individual model view has a chart showing false positive and false negative rates (defined in `PerformancePlot.tsx`). The axis of this plot was in percent, but the annotations on the bars themselves were decimal. This problem was due to the incorrect key to `PerformanceMetrics.ts` being passed into `FormatMetrics.ts` by `PerformancePlot.tsx` - in the `PerformanceMetric` map, the names `fallout_rate` and `miss_rate` are used. This fixes that mismatch.

Had to force in because for some reason the successful builds were not registered.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>

* update notice file (#286)

Signed-off-by: Roman Lutz <rolutz@microsoft.com>

* remove ipywidgets requirement and also duplicate (unused) gevent (#287)

Signed-off-by: Roman Lutz <rolutz@microsoft.com>

* move footer up in axis config dialog, add legend text, fix indexing for (#288)

what if and inspection view, add big data support, add map shift dialog

* Add localisation files to prettierignore (#290)

The language specific files `en.*.json` are produced by the localisation builds, and not by us. Remove them from prettier.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>

* fix label alignment, fix what if for categoricals, change index to come (#292)

from explanation dataset

* Minor style fix (#289)

* style fix

* fabric build-in

* notebook adjustments to make them work with latest changes

Signed-off-by: Roman Lutz <rolutz@microsoft.com>

* remove debugging statement

Signed-off-by: Roman Lutz <rolutz@microsoft.com>

* yarn lintfix ignore README

Signed-off-by: Roman Lutz <rolutz@microsoft.com>

Co-authored-by: Roman Lutz <rolutz@microsoft.com>
Co-authored-by: Richard Edgar <riedgar@microsoft.com>
Co-authored-by: Ilya Matiach <ilmat@microsoft.com>
Co-authored-by: Bo <71688188+zhb000@users.noreply.github.com>