Fix issue with missing directories (#1085)
When saving to AzureML, directories are only implicit. This means that empty directories don't exist, which is a problem when we download an `RAIInsights` directory where one or more of the tools wasn't run.

This change tweaks the causal, counterfactual and error analysis managers to be tolerant of this. It is not applied to the explainer manager because that does create files in its manager directory even if no explanation is produced.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>